Rewrite TSR installation logic to be cleaner
Previously, the handoff between the code that fills the buffer and the
code that finalizes the installation was pretty messy, with calculations
being done in awkward places that made it harder to extend. This commit
streamlines things by encoding most of what the finalizer needs to know
in how the data is laid out in the buffer.