Convert bundle values to wstrings at parse time
Bundle values are wstrings of the form "KEY=VALUE". Previously, when
parsing a key-value pair, we would save the value by saving a pointer
into the middle of the wstring data. This pointer would not be a
wstring itself.

This commit mutates the data at parse time, so that the parsed values
are proper wstrings. The keys are partially overwritten, which is a
little ugly, but we don't need to read them again.