Fix various bugs, get palettes/fonts working again
Also in this commit: implement a proper preview mode which uses the
exact same video code as the resident handler does. I first did this
to make it easier to debug the video code, but it also has the benefit
of making sure the preview is accurate -- for example, the preview
didn't set fonts before, but now it does.