Don't allow unknown keys in the data bundle
Previously, if the bundled data contained lots of zero bytes, it was
surprisingly easy for the validation code to miss malformed data. This
was because invalid keys were ignored: the parsing code would ignore
most of the data, until it came to a double-zero, at which point it
would stop parsing altogether.

This commit addresses this by disallowing invalid keys in the bundle.