Change bundle format to avoid splitting on '='
Previously, the bundle was a list of strings of the form "KEY=VALUE",
which was awkward to work with: it required string manipulation in
assembly, and the bundle parser mutated those strings in place due to
considerations of keeping memory usage under 64k.

This commit changes the bundle format such that every key-value pair
consists of two separate strings: one for the key, and one for the
value. This simplifies the code, and the "new" subcommand magically
produces valid bundles now.